% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth}
\alias{auth}
\title{Get Lemon Markets Bearer Token}
\usage{
auth(
  client_id = Sys.getenv("LEMON_MARKETS_CLIENT_ID"),
  client_secret = Sys.getenv("LEMON_MARKETS_CLIENT_SECRET"),
  renew = FALSE
)
}
\arguments{
\item{client_id}{character; Client ID to be used.}

\item{client_secret}{character; Client Secret to be used.}

\item{renew}{logical; Specify if token shall be renewed, even if already existing.}
}
\description{
Retrive a bearer/session token to access the Lemonmarkets
API.
}
