% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{balance}
\alias{balance}
\title{Get Balance of Account or Space}
\usage{
balance(space_id = NULL)
}
\arguments{
\item{space_id}{character; Optional space ID for which balance shall be retrieved. See also \code{list_spaces()}}
}
\description{
Get Balance of Account or Space
}
\examples{
\dontrun{
balance()
balance(space_id = "e44907b7-d131-4ec9-9647-a2649480003d")
}
}
