% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_order.R
\name{create_order}
\alias{create_order}
\title{Create Order}
\usage{
create_order(
  space_id,
  isin,
  quantity,
  valid_until = Sys.time() + 86400,
  activate = TRUE
)
}
\arguments{
\item{space_id}{character; Space ID for which balance shall be retrieved. See also \code{list_spaces()}}

\item{isin}{character; ISIN of instrument to be retrieved.}

\item{quantity}{integer; Number of stocks to be bought/sold.}

\item{valid_until}{POSIXct; Time until order shall be valid.}

\item{activate}{logical; Automatically activate order.}
}
\description{
Create Order
}
\examples{
\dontrun{
  create_order(space_id = "e44907b7-d131-4ec9-9647-a2649480003d",
               isin = "DE0008404005",
               quantity = 1)
}
}
