% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ohcl_data.R
\name{get_ohcl_data}
\alias{get_ohcl_data}
\title{Get Market data in OHCL Format}
\usage{
get_ohcl_data(
  isin,
  freq = "daily",
  interval = 1,
  from = Sys.Date() - 30,
  is_live = Sys.getenv("LEMON_MARKETS_IS_LIVE", FALSE)
)
}
\arguments{
\item{isin}{character; ISIN of instrument to be retrieved.}

\item{freq}{character; Frequency of time series to be retrieved, one of 'daily', 'hourly', 'min'}

\item{interval}{character; Interval of time series, defaults to 1. E.g. freq = "hourly" and interval = 2 means every to hours.}

\item{from}{character; Starting date to be retrieved in ISO format YYYY-mm-dd}

\item{is_live}{logical; Specify if data through live API shall be retrieved.}
}
\description{
Get Market data in OHCL Format
}
