% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orders.R
\name{get_orders}
\alias{get_orders}
\title{Get Orders of Space}
\usage{
get_orders(space_id)
}
\arguments{
\item{space_id}{character; Space ID for which balance shall be retrieved. See also \code{list_spaces()}}
}
\value{
data.frame including orders
}
\description{
Get Orders of Space
}
\examples{
\dontrun{
get_orders(space_id = "e44907b7-d131-4ec9-9647-a2649480003d")
}
}
