% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_portfolio.R
\name{get_portfolio}
\alias{get_portfolio}
\title{Get Portfolio Positions of Space}
\usage{
get_portfolio(space_id)
}
\arguments{
\item{space_id}{character; Space ID for which balance shall be retrieved. See also \code{list_spaces()}}
}
\value{
data.frame including porfolio positions.
}
\description{
Get Portfolio Positions of Space
}
\examples{
\dontrun{
get_portfolio(space_id = "e44907b7-d131-4ec9-9647-a2649480003d")
}
}
