% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Get Lemon Markets Bearer Token}
\usage{
get_token(
  client_id = Sys.getenv("LEMON_MARKETS_CLIENT_ID"),
  client_secret = Sys.getenv("LEMON_MARKETS_CLIENT_SECRET")
)
}
\arguments{
\item{client_id}{character; Client ID to be used.}

\item{client_secret}{character; Client Secret to be used.}
}
\description{
Retrive a bearer/session token to access the Lemonmarkets
API.
}
