% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_components.R
\name{index_components}
\alias{index_components}
\title{Get All components from a specified Index}
\usage{
index_components(index_name = "DAX")
}
\arguments{
\item{index_name}{character; Specified Index.}
}
\description{
Currently, only the DAX index is supported
}
