% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohcl.R
\name{ohcl}
\alias{ohcl}
\title{Get Market data in OHCL Format}
\usage{
ohcl(isin, freq = "daily", interval = 1, from = Sys.Date() - 30)
}
\arguments{
\item{isin}{character; ISIN of instrument to be retrieved.}

\item{freq}{character; Frequency of time series to be retrieved, one of 'daily', 'hourly', 'min'}

\item{interval}{character; Interval of time series, defaults to 1. E.g. freq = "hourly" and interval = 2 means every to hours.}

\item{from}{character; Starting date to be retrieved in ISO format YYYY-mm-dd}
}
\value{
zoo objects with the columns "open", "high", "low" and "close"
}
\description{
Get Market data in OHCL Format
}
\examples{
# Get Data for Deutsche Post
ohcl("DE0005552004")
}
