% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{create_order}
\alias{create_order}
\alias{get_orders}
\alias{delete_order}
\alias{activate_order}
\title{Create Order}
\usage{
create_order(
  space_id,
  isin,
  quantity,
  valid_until = Sys.time() + 86400,
  activate = TRUE
)

get_orders(space_id)

delete_order(space_id, order_id)

activate_order(space_id, order_id, verbose = getOption("verbose"))
}
\arguments{
\item{space_id}{character; Space ID for which balance shall be retrieved. See also \code{list_spaces()}}

\item{isin}{character; ISIN of instrument to be retrieved.}

\item{quantity}{integer; Number of stocks to be bought/sold.}

\item{valid_until}{POSIXct; Time until order shall be valid.}

\item{activate}{logical; Automatically activate order.}

\item{order_id}{character; UUID of order to be deleted.}

\item{verbose}{logical; Specify if output shall be printed to the console.}
}
\value{
data.frame including orders
}
\description{
Create Order

Get Orders of Space
}
\examples{
\dontrun{
  create_order(space_id = "e44907b7-d131-4ec9-9647-a2649480003d",
               isin = "DE0008404005",
               quantity = 1)
}
\dontrun{
get_orders(space_id = "e44907b7-d131-4ec9-9647-a2649480003d")
}
}
