% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{balance}
\alias{balance}
\title{Get Balance of Account or Space}
\usage{
balance(space_id = NULL)
}
\description{
Get Balance of Account or Space
}
\examples{
\dontrun{
balance()
balance(space_id = "69fc7fbe-802d-4cd4-9f02-8965536c6896")
}
}
