% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments_search.R
\name{instruments_search}
\alias{instruments_search}
\title{Retrieve tradable instruments}
\usage{
instruments_search(search = NULL, isin = NULL)
}
\arguments{
\item{search}{character; Search term to find requested instrument.}

\item{isin}{character; ISIN of instrument to be retrieved.}
}
\description{
Retrieve tradable instruments
}
\examples{
\dontrun{
  instruments_search(search = "daimler")
}
}
